+++ behavior_machine/core/state_status.py
@@ -1,6 +1,5 @@
 import enum
 
-
 @enum.unique
 class StateStatus(enum.Enum):
     UNKNOWN = -1     # Unknown