+++ behavior_machine/library/__init__.py
@@ -3,3 +3,4 @@ from .sequential_state import SequentialState
 from .common_state import IdleState, WaitState
 from .parallel_state import ParallelState
 from .selector_state import SelectorState
+from .common_state import SetFlowState, SetFlowFromBoardState, SaveFlowState